Add files via upload
Separate files for easier future development of the GUI ddcutil program. Having problems with my MSI Strix MAG 251RX VCP v2.1 communication protocol as it is "unreliable" and sometimes it works on the first time like 20% of the time. The other times I need to re-initiate the ddcutil command from the terminal a couple of times for it to work (sometimes the terminal outputs error, but indeed the brightness on the monitor changes). I checked VCP documentation and some v2.1 of Dell work flawlessly, others don't.
My Philips PHL 243V7 with VCP v2.2 work flawlessly.

Must fix the GUI to initiate the "ddcutil list" command a couple of times before it captures it in the GUI so that it can show the correct monitor. Now whenever I get the 3rd MSI monitor is it shows as 3rd Philips one.